moved Py class into its own file (#1649)
